cleanup and add welcome screen
